modularised normal functions
