--- test_files/rules.yml
@@ -1,14 +1,0 @@
-- rules:
-    - attribute: "Title"
-      value: "Push Event"
-      op: "eq"
-  transformation:
-    Title: "PUUUUUUUUUSH Event"
-    color: "255,0,0"
-- rules:
-    - attribute: "Title"
-      value: "Pipeline Event"
-      op: "eq"
-  transformation:
-    Title: "PIIIIIPE Event"
-    color: "0,255,0"
